-separated page for the color palette -removed all copy to clipboard functionality from the colors guide
