+++ docs/src/routes.tsx
@@ -6,6 +6,7 @@ import DocsLayout from './components/DocsLayout'
 import DocsRoot from './components/DocsRoot'
 
 import Accessibility from './views/Accessibility'
+import Colors from './views/Colors'
 import ColorPalette from './views/ColorPalette'
 import ColorSchemes from './views/ColorSchemes'
 
@@ -133,6 +134,7 @@ const Router = () => (
           path="/integrate-custom-components"
           component={IntegrateCustomComponents}
         />
+        <DocsLayout exact path="/colors" component={Colors} />
         <DocsLayout exact path="/color-palette" component={ColorPalette} />
         <DocsLayout exact path="/color-schemes" component={ColorSchemes} />
         <DocsLayout exact path="/*" component={PageNotFound} />
