+++ docs/src/views/Layout.tsx
@@ -209,7 +209,7 @@ export default () => (
       <strong>Each of these examples is optimized in terms of DOM elements rendered</strong>.
     </blockquote>
     <GuidesNavigationFooter
-      previous={{ name: 'Colors', url: 'color-palette' }}
+      previous={{ name: 'Colors', url: 'colors' }}
       next={{ name: 'Integrate Custom Components', url: 'integrate-custom-components' }}
     />
   </DocPage>
